Include resolved net-config in VM status
This patch updates the VM API to have status.network.config, a
field for including the resolved, intended network configuration
for a VM. This is useful for when a user is deploying a VM for
which there is no bootstrap engine. This new field will surface
the *intended* network configuration for the VM, including the
possibly IPAM-derived IP address(es) and DNS information.