Remove need for VM Image in Update path
Post create, I'd like us to never have to depend in the image resource
existing since it may get deleted. Fix the two situations where the
image was being used.

First the JSON_EXTRA_CONFIG contains RUN.container that uses a template
to fill in the image name so vcsim creates a container under the covers.
This was used for old local gce2e testing, and isn't used by us but some
other teams have a desire to use it so keep it around. Now that is just
set on create in the EC and we don't need to re-eval it later on.

Second remove the V1Alpha1Compatible condition check and always check
the EC. This does change the behavior that if an image did not have the
condition but had guestinfo.vmservice.defer-cloud-init=ready, we'll now
always set it to enabled. I don't think this will really matter.