UpdateExtraConfig for powered on VMs, iter list once
This patch modifies the logic for updating ExtraConfig to
support powered on VMs. This ensures powered on VMs that
have vGPUs will have the flag in ExtraConfig removed that
DRS queries to determine whether or not to power off the
VM during vMotion.

This patch also optimizes the behavior of the UpdateEC
function to only iterate over the existing EC list once.