ConfigSpec by value
This patch updates the way the VM ConfigSpec is handled
by several util functions and the vSphere provider to
be copy-by-value instead of as a pointer. This is because
Golang is much more efficient at handling memory on the
stack versus the heap.

The primary reason for this patch is to change the
perception of the configSpec variable throughout the rest
of its call stack. The spec is not optional, and moving
away from a pointer reinforces that notion.