Introduce vm.spec.image reference
This patch introduces the field spec.image to the VirtualMachine API.
This new field allows users to explicitly specify the reference to the
VirtualMachineImage or ClusterVirtualMachineImage resource from which
a VM is deployed. Additionally, the existing field spec.imageName is
no longer mutated in place if the value of the field is an image's
display name. Instead, the resolved image reference is mutated into
spec.image.

This patch also removes status.image from the VirtualMachine API
since it is now redundant.