Combine GH actions for test/coverage
This patch removes the distinct GitHub action for parsing the
code coverage produced by the test action. Instead, the coverage
logic is now handled at the end of the test action. Previously
the coverage logic needed to be a distinct step since there were
two, separate test jobs. However, now those are combined into one,
meaning we can remove the need to upload the coverage artifact to
be processed later.