GuestInfo prober fetches required props only
This patch updates the GuestInfo prober to only fetch the properties
that are part of the specified probe definition. Previously the VM's
entire config.extraConfig property path was retrieved, resulting in
a very expensive operation, depending on the size of the data in that
property. By focusing on specific properties, the number of bytes
transferred is limited to the size of the property values required to
evaluate the probe. This change is possible because vSphere supports
fetching property paths with indexed values. For example, instead of
fetching all of config.extraConfig, an array, just to access the element
with the key "hello", it is possible to fetch the property path
config.extraConfig["hello"].

Additionally, this patch forgoes the use of the GoVmomi VM helper method
"Properties" for retrieving the values. GoVmomi uses Golang reflection
to unmarshal the result of a property retrieval, and reflection is quite
expensive. Instead, the GuestInfo prober inspects the results of the
retrieval directly, without unmarshaling the data into a
mo.VirtualMachine.