Invert kubebuilder annotations
This patch inverts the kubebuilder annotations for the
v1alpha3 schema. The value of this is that documentation
generated from the GoDocs no longer contends with the
kubebuilder comments. For example, this is true for the
results of "make generate-api-docs" as well as the API
documentation hosted at https://pkg.go.dev. To validate
the latter locally, install the pkgsite command with
go install golang.org/x/pkgsite/cmd/pkgsite@latest and
run "pkgsite ." from the root of this repository.

This patch also updates the tool used to generate docs
from the CRDs, crd-ref-docs, to v0.0.12.