Store resolved Cloud-Init instance ID in spec
This patch stores the Cloud-Init instance ID derived from
either spec, the object UID, or the annotation, back into
spec if the derived version is different than what is in
spec. Additionally, the annotation that can overrride the
instance ID is deleted once its value is stored in the
spec.

This patch also updates some documentation to make the
logic more clear, as well as provide historical context
for why certain logic exists in the first place.