Honor v1a1 VMMetadata ConfigMapName in v1a3 using annotations (#498)
Our initial implementation involved a configMap fallback when the secret for the VM was not
present. This can cause issues when a secret with the same name is present in the same
namespace. Therefore, this change populates an annotation during up conversion to
handle cases where v1a3 needs to use a configMap instead of a secret. On down conversion back
to v1a1, the ConfigMapName field is set based on existance of the same annotation.