Add last VM Resize annotation
This annotation will be used for VM resizing to keep track of which
and at what revision a VM was last resize from.

Add very basic Set/Get methods for this. This is really just to
get what & how is stored in the annotation out there. We'll add
matcher helpers on top of these later.