Plumb through basic end to end resize
This is just the minimal plumbing needed to go through resize when the
VM is in steady-state off. Lots of supporting plumbing is still needed
like handling the last resize annotation when the VM Class does not
exist, and building the full ConfigSpec that includes expected network
devices, but this lets us work on the end to end resize.

This also copies the existing getUpdateArgsFn hack that we have as a
holdover from the old Session days that we'll likely need to clean up
as we have time.