Support system:masters for IsPrivilegedUser
This patch updates the IsPrivilegedUser function for webhooks
to support the membership in the system:masters group as signal
the user is privileged. This is consistent with the Kubernetes
documentation at
https://kubernetes.io/docs/reference/access-authn-authz/rbac/#user-facing-roles.

This also means the integration tests in this project that
depend on envtest will now access the API server as an admin.