Protect VirtualMachine spec.biosUUID
This patch protects the VirtualMachine field spec.biosUUID from
being set when a new VM is created unless by a privileged user.
This logic resides in the VM's mutation webhook instead of the
validation webhook, since the former also defaults the value,
and the latter would have no way of knowing if the default
value originated from the instigating user or the mutation
webhook.