Import net-operator-api instead of vendoring
This patch updates VM Op to import the net-operator API types
from github.com/vmware-tanzu/net-operator-api instead of vendoring
the types. The CRD for the net-operator NetworkInterface API is
still vendored as it is required for testing. The CRD for the
Vmxnet3NetworkInterface is still vendored as well since it used
to be vendored. However, this API is not used by VM Operator and
its CRD will likely be removed from vendoring in the future.