Ginkgo v2.19.0
This patch updates ginkgo in the main module from v2.17.1 to
v2.19.0 and in hack/tools from v2.15.0 to v2.19.0. This is to
ensure both versions are the same. When they are not, an error
is received in GitHub actions like so:

    Ginkgo detected a version mismatch between the Ginkgo CLI
    and the version of Ginkgo imported by your packages:
      Ginkgo CLI Version:
        2.15.0
      Mismatched package versions found:
        2.17.1

In order to keep the versions aligned, they were both updated
to the most recent release of ginkgo.