Rm k8s.io/utils, add pkg/util/ptr, use net.ParseCIDR
This patch removes the direct dependency on k8s.io/utils. This
module was used for its k8s.io/utils/ptr and k8s.io/utils/net
packages. The former has been replaced with this repository's
own pkg/util/ptr package. The latter is no longer required. It
was used for the function ParseIPNets, which is just a wrapper
around the stdlib Golang net.ParseCIDR function with support for
IP4 addresses with leading zeros (removed in Go 1.17), ex.
"001.001.001.001". While this is a legal representation of an IP
address, it is not possible for these ranges to exist in VM Service
due to Supervisor using Golang's standard net.ParseCIDR function
for validating incoming IP4 network ranges.

This patch further simplifies VM Operator's dependency graph.