Exclude ./api from coverage
This patch excludes the ./api module from coverage. This is
because there is too much generated code in the API schemas
that is not able to be tested due to the way the code is
generated. For example, in ./api/v1alpha2/zz_generated.conversions.go:

    func RegisterConversions(s *runtime.Scheme) error {
        if err := s.AddGeneratedConversionFunc((*ClusterVirtualMachineImage)(nil), (*v1alpha3.ClusterVirtualMachineImage)(nil), func(a, b interface{}, scope conversion.Scope) error {
            return Convert_v1alpha2_ClusterVirtualMachineImage_To_v1alpha3_ClusterVirtualMachineImage(a.(*ClusterVirtualMachineImage), b.(*v1alpha3.ClusterVirtualMachineImage), scope)
        }); err != nil {
            return err
        }

There are hundreds of these in the v1alpha1, v1alpha2, and v1alpha3
schema packages, and for each one, there is no way to cover the
logic where an error is returned due to the generated code.

Instead of having the API module cause the project's overall
coverage metric dip into less-than-ideal numbers, this patch
excludes the API module from coverage. Please note, the API
module is *still* tested, and the "test" and "test-nocover"
targets *will* fail if the API tests fail. The only difference
is that the API module is no longer counted towards overall
code coverage for the project.