Replace gopkg.in/yaml.v3 with sigs.k8s.io/yaml
This patch replaces all uses of gopkg.in/yaml.v3 with sigs.k8s.io/yaml.
This change is to ensure there is a singular convention for how objects
are marshaled/unmarshled to/from YAML within VM Operator, regardless of
whether they are Kubernetes runtime objects. The package
sigs.k8s.io/yaml is in fact just a wrapper around gopkg.in/yaml, with
the additional logic of first marshaling the objects to JSON before
converting them to YAML.