Fix coverage for both patch packages
This patch fixes an issue in the "patch" packages where they would
occasionally report only 16% coverage instead of 78%. For example,
https://github.com/vmware-tanzu/vm-operator/pull/553 showed pkg/patch
at 78%, whereas
https://github.com/vmware-tanzu/vm-operator/pull/552 showed it at 16%.
When the lower coverage was detected, Ginkgo did not detect any tests
in the package, ex.:

        Running Suite: Patch Suite - /home/runner/work/vm-operator/vm-operator/pkg/patch
        ================================================================================
        Random Seed: 1717175300

        Will run 0 of 0 specs

        Ran 0 of 0 Specs in 0.000 seconds
        SUCCESS! -- 0 Passed | 0 Failed | 0 Pending | 0 Skipped

This was RCAd to be the mixing of the pure-Golang style testing
versus Ginkgo. This patch updates both packags to use Ginkgo.