Fix VM controller VM Class reconcile request
An empty VMClass.Spec.ControllerName name would not cause us to
reconcile the VMs when its VM Class was changed.

Fix requeueDelay() to only fall into the IP retry if the VM is expected
to have networking in the first place.

Use a valid UUID for dummy InstanceUUID. The prior value along with the
schema upgrade would fail due to field validation. Due to how the patch
helpers breaks up the patches, this was causing an error retry instead
of the expected requeueDelay() usage.

Remove a bogus Creating Phase test: the controller used to set the Phase
in some cases, but Phase is gone and the controller doesn't set
Conditions directly. This test wasn't really testing anything.

Fix a race when testing VM deletion so we actually assert the VM was
deleted.