Resize Brownfield VMs
VMs prior to this feature will not have the last-resize annotation that
we will set on new VMs at create time. When the ClassName changes set
the last-resized annotation so we'll evaluate a resize when the VM
powerstate allows.

Add a quick for a braino and last second refactor that caused us to
match all devices instead of just PCI passthru devices. I need to do
some more clean up and refactoring of the tests to make this more
natural to test.