Fix v1a1 CloudInit with no user-data only ssh public keys
The v1a1 VM Metadata (Bootstrap) didn't have a field to specify the CM
or Secret key for the CloudInit userdata. In v1a2, we added a Key field,
and that field is required so on version convert we use "user-data".
However, that key - or any user data - may not exist in the CM or Secret
and the user may only want to provide the default user SSH public keys,
allow that to succeed.