Revert to returning early if the pause annotation is present on the VM (#620)
In the recent changes to support the break-glass admin overrides, we
changed the behavior of the pause annotation for regular (non delete)
reconciles so that we now _always_ add the finalizer regardless of the
pause annotation.

This change fixes this by reverting the old behavior of returning
early if the pause annotation is present.  This also means that now
both reconcile methods (Create/Update and Delete) behave consistently
when this annotation is present.