Introduce an annotation for unconditional opt into backup
Today, we have certain exclusion rules in place that govern whether or
not we persist the configuration of VirtualMachine in the ExtraConfig
for use by backup/restore or disaster recovery solutions.  However,
there might be cases where an entity wants to explicitly opt into
backup.

This change introduces an annotation that can be placed on the
VirtualMachine resource that lets VM operator skip the exclusion
checks when determining whether or not to store the VM's config in
ExtraConfig.