Introduce the ExtraConfig key to allow opt-out from auto register (#631)
Currently, we will unconditionally try to register a VM when it is
restored using a backup/restore vendor, or a DR solution.  This
isn't always ideal if the solution wants to prepare the infra, or
customize the VM on vSphere before registering the VM with Supervisor.

This change introduces an ExtraConfig key that can be used to opt-out
of the automatic registration workflow.