Rename the pause-reconcile annotation to paused
This change renames the "pause-reconcile" annotation used to pause the
reconciliation of a VM to "paused".  This is consistent with other
components using "paused" to indicate an entity being paused.

Additionally, introduce a helper package that can be used to check if
an annotation is present (regardless of the value) in a readable way.