resize: handle CPU and memory reservations/limits for limited resize
We want to update the VM reservations and limits as the VM is resized.
Note that the unspecified in the class we have to explicitly default to
Reservation:0,Limit:-1 (best effort) in our VM creation ConfigSpec so
that will update the values, for example when the VM is going from a
guaranteed class to a best-effort one.

Remove the old UpdateConfigSpecCPUAllocation and
UpdateConfigSpecMemoryAllocation functions. These were no longer being
called, and not really correct since it didn't take the class ConfigSpec
into account. Instead, we can use the new compare functions wrote for
full resize.