Avoid duplicated VM ReconcileNormal events
Only emit a default event if ReconcileNormal fails and it was neither
a create nor update op.

While here, fix a long standing bug if that we should update the
probe manager status even if CreateOrUpdate fails. If the VM
Spec.ReadinessProbe has been changed we should update what's
being probed even if we reconcile fails for some reason.