Start infra capability controller for non-leader pods too
We want capabilities updates to be reflected in the non-leader pod as
well so the webhooks are aware of the current enabled capabilities.

Use the ConfigMap Cache we create as the k8s Client instead of the mgr
Client since the CM will be in the cache, where as the mgr Client does
not cache CM (and Secrets)

Remove the VMProvider from the infra capability controller. It was not
used, and we don't want it on the non-leader pods due to VC session
limits (and it would need the other infra controllers to be non-leader
as well to pick up cred rotations)