Controller-runtime v0.19.0 & tools v0.16.1
This patch updates VM Operator to Controller-Runtime v0.19.0 and
Controller-Tools v0.16.1 as it is required to work with the new
k8s.io dependency pulled in by Controller-Runtime.

This patch also addresses a bunch of linting issues brought to
light by stale linting caches on developer systems and on GitHub
actions where part of the cache key is the sum of the go.mod file.
Since the module file changed, it ran with a clean cache, revelealing
the issues.

Finally, this patch updates a few tests to set conditions' field
LastTransitionTime to a non-empty value since that is now a required
field.