Use ResourcePool field out of the VM properties
We can save one RT in every reconcile since we have to fetch
the VM properties anyways until we can really fix this old
session code.

Update the session tests to use the same properties selector
that is used in the prod code. For the short term update the
sessions tests to use Summary.Runtime instead of Runtime for
the power state. Longer term it would be better for these tests
to not conflate the selector used for UpdateVirtualMachine()
input and test assertions.