Allow CAPV to still create VMs in draining zones
CAPV needs to be able to provision VMs in draining zones prior to the
zone being deleted.

Since TKG is now managed via kapp it does not have a well-known NS so
we cannot hard code the service account name like we do with other
accounts. This check does not need to be fool proof in the sense that
this is advisory and everything in the zone is eventually going to be
deleted.