Strip managed fields for cached objects
This patch adds support for stripping managed fields for objects
entering the cache. The controller-runtime documentation for this
transformer states:

    If you are not explicitly accessing managedFields from your code,
    setting this as `DefaultTransform` on the cache can lead to
    significant reduction in memory usage.

Given that VM Operator never directly accesses the managed fields
for any object, there is no reason we should not be using this
transformer.

Please see
https://github.com/kubernetes-sigs/controller-runtime/pull/2791
for more information on this feature.