Handle unbound PVCs Immediate SC during placement
We must wait for PVCs with an immediate StorageClass to be bound,
so we'll know the accessible zones. Similarly, for bound PVCs we
should get its accessible zones as a part of the placement constraints.

This is first pass intended to address the lowest hanging PVC
placement work, with a lot of reworking of placement to be done
later. Specifically, we should be smarter about whether we'll
actually do placement or not, based on the cluster and VM config.