Support unencrypted VMs w default provider
This patch adds support for deploying unencrypted VMs even when
there is a default key provider present on the system. This is
a change from the previous behavior that would disallow deploying
a VM on a non-encryption storage class or sans vTPM if there was
a default key provider present. This would in effect enforce the
need for encryption and its requirements. This patch treats new
VMs as special and allows it to skip encryption IFF:

- the discovered provider is the default
- there is no vTPM
- the storage class does not support encryption