Fix validate classless vm on update
This patch fixes the validation webhook for updating classless VMs.
The enabling of the Resize CPU FSS caused this validation to fail
for importing classless VMs since the VM controller uses the deferred
patch helper. The patch helper was failing to patch VMs with no class
since the validation webhook prevented patches with empty class names.