Disallow PVCs with WaitForFirstConsumer StorageClass
CSI does not support this for VMs at the moment since it does not wait
for VMs to be placed. CSI doesn't produce a very helpful error message
for this condition, and there is an ask to provide immediate feedback
so put this best-effort check in the validation webhook.

This uses field.Forbidden instead of field.NotSupported since the later
does not take a string to provide a useful error message.