Fix how crypto detects device sans policy
This patch fixes how the crypto logic detects whether or not a
device is being added/edited sans policy that supports encryption.

Previously this happened via a direct IOFILTERS check, but that may
not be set in the profile on the device. Instead the profile ID must
be used to query whether or not the profile supports encryption.