Aggregate CreateOrPatch errors during StoragePolicyQuota reconcile
If there is some create or patch error for one StoragePolicyUsage
don't return early from the reconcile: the others may succeed.
This was prompted by the double take of this function ending with
"return err" where by the end err is always nil.

While here, add missing assertion for the expected CABundle.