Support building cross-arch images
This patch adds support for building images targeting an
architecture different than the host's by specifying
GOARCH with the desired architecture. For example, the
following command will build a linux/amd64 image, even
when executed on an Apple silicon mac:

    GOARCH=amd64 make image-build

Please note, this feature requires either Podman or a
version of Docker that supports buildx.

This patch also:

- Renames all the "docker-*" targets to "image-*". The docker
  aliases are retained for backwards-compatibility.

- Saves any images built with the "image-build" target to the
  file "artifacts/vmoperator-controller-GOOS_GOARCH.tar.