Unifiy build process up and downstream
This patch changes the upstream image-build target and Dockerfile
to no longer do a staged build, building VM Op inside of the
container. Instead, the binaries are now built prior to building
the container.

This, as well as a few other changes to the version variables in
the Makefile, allow the up and downstream build processes for VM
Op to be unified. Now the internal process can invoke the following
to build VM Op from the upstream Makefile:

    ADDITIONAL_CRI_BUILD_FLAGS="--network=host" \
    BASE_IMAGE="<INTERNAL_CONTAINER_IMAGE>" \
    BUILD_BRANCH="$(BUILD_BRANCH)" \
    BUILD_COMMIT="$(BUILD_COMMIT)" \
    BUILD_NUMBER="$(BUILD_NUMBER)" \
    BUILD_VERSION="$(BUILD_VERSION)" \
    IMAGE="$(IMAGE)" \
    IMAGE_TAG="$(IMAGE_TAG)" \
    IMAGE_VERSION="$(IMAGE_VERSION)" \
    PRDCT_VERSION="$(PRDCT_VERSION)" \
    make image-build-amd64