Minimum version selection
This patch updates all of the go.mod files in this repository,
*other than the one for building the manager binary*, to specify
the minimum required version of Go and minimum required versions
for dependencies.

This change allows downstream consumers of things like VM Op's API
module to use a version of Go older than the one used by this repo.
It turns out that Go 1.21 changed the meaning of the Go version in
go.mod. Prior to 1.21, the version of Go in go.mod was informational.
With Go 1.21 and after, it's required. That means if someone depends
on VM Op's API module, and the module specifies Go 1.23.2, then the
project depending on us requires 1.23.2.

Of course our main binary still specifies the latest Go version and
newest dependencies for handling CVEs. However, our sub-modules that
are made to be pulled into *other* targets, such as our own manager,
have no reason to specify the latest version of Go or dependencies.