Set CPU/mem limit to -1 if reservation & no limit
This patch addresses an issue with vSphere placement APIs that expect
the optional Limit field for CPU/Memory allocation to be set to -1 if
there is a requested reservation and no limit. vSphere *should* be fine
if the Limit field is omitted since it is optional, but while that is
being chased down, this work-around addresses the issue.