OVF Cache Package
This patch refactors the OVF cache into an context-based
package so the cache may be shared across the project and
not limited to the vSphere provider. This also simplifies
a future refactor to transition the cache to using Secret/
ConfigMap resources in Kube for caching the OVF envelopes
as opposed to an in-memory store.