Ignore async signal for deleted VMs
This patch fixes a bug that results in an infinite loop of calls
to the vSphere VM's Delete API. The vSphere VM Delete API invokes
Reload on the VM, causing all property collectors to emit MODIFY
events for the VM. The VM Op async watcher sees this event and thus
enqueues the VM for reconcile. The reconciler sees the VM has a
non-zero deletion timestamp and calls the vSphere VM's Delete API.
Rinse and repeat.

Normally this would not be an issue, as the vSphere VM *should* be
deleted, causing the loop to stop. However, when a bug occurs in
vpxd on vSphere that results in a deployed VM in an invalid state,
it cannot be successfully deleted. The result is VM Op spewing
dozens of Delete calls a second for the VM due to the async signal
received from the property collector, caused by the Reload API,
invoked by the Delete API.

This patch fixes the issue by no longer enqueuing a reconcile for
VMs that have non-zero deletion timestamps, i.e. they are being
deleted.